+++ pdfalyzer/decorators/pdf_object_properties.py
@@ -5,12 +5,14 @@ from pypdf.errors import PdfReadError
 from pypdf.generic import DictionaryObject, IndirectObject, NullObject, NumberObject, PdfObject
 from rich.text import Text
 
-from pdfalyzer.helpers.pdf_object_helper import pypdf_class_name
-from pdfalyzer.helpers.rich_text_helper import comma_join_txt
-from pdfalyzer.helpers.string_helper import coerce_address, is_array_idx, props_string_indented, root_address
-from pdfalyzer.output.styles.node_colors import get_class_style, get_class_style_dim, get_class_style_italic, get_label_style
+from pdfalyzer.output.highlighter import PdfHighlighter
+from pdfalyzer.output.theme import (COMPLETE_THEME_DICT, DEFAULT_LABEL_STYLE, get_class_style,
+     get_class_style_dim, get_class_style_italic)
 from pdfalyzer.util.adobe_strings import *
-from pdfalyzer.util.logging import log, log_console, log_highlighter, log_trace
+from pdfalyzer.util.helpers.pdf_object_helper import pypdf_class_name
+from pdfalyzer.util.helpers.rich_text_helper import comma_join_txt
+from pdfalyzer.util.helpers.string_helper import coerce_address, is_array_idx, props_string_indented, root_address
+from pdfalyzer.util.logging import highlight, log, log_console, log_highlighter, log_trace, pdf_highlighter
 
 
 @dataclass
@@ -40,6 +42,16 @@ class PdfObjectProperties:
     @property
     def type(self) -> str | None:
         return self._type
+
+    @property
+    def label_style(self) -> str:
+        type_no_slash = (self.type or '').removeprefix('/')
+        sub_type = self.sub_type or ''
+
+        if sub_type.startswith(GO_TO_R) or sub_type.startswith(GO_TO_E):
+            return COMPLETE_THEME_DICT[GO_TO_R]
+
+        return COMPLETE_THEME_DICT.get(PdfHighlighter.prefixed_style(type_no_slash), DEFAULT_LABEL_STYLE)
 
     @type.setter
     def type(self, _type: str | None):
@@ -97,7 +109,7 @@ class PdfObjectProperties:
             if isinstance(reference_key, int):
                 key_style = 'grey'
             else:
-                key_style = log_highlighter.get_style(reference_key)
+                key_style = pdf_highlighter.get_style(reference_key) or log_highlighter.get_style(reference_key)
 
         with_resolved_refs = self._resolve_references(reference_key, row_obj, pdfalyzer)
         value_style = key_style if isinstance(row_obj, str) else get_class_style(row_obj)
@@ -112,7 +124,7 @@ class PdfObjectProperties:
         text = Text('<', style='white')
         text.append(f'{self.idnum}', style='bright_white')
         text.append(':', style='white')
-        text.append(self.label[1:], style=f"{get_label_style(self.label)} {'underline' if underline else ''} bold")
+        text.append(self.label[1:], style=f"bold {self.label_style} {'underline' if underline else ''}")
         text.append('(', style='white')
         text.append(pypdf_class_name(self.obj), style=get_class_style_italic(self.obj))
         text.append(')', style='white')
@@ -153,7 +165,7 @@ class PdfObjectProperties:
             return cls.__rich_without_underline__(obj)
         elif isinstance(obj, str):
             if 'http' in obj or obj.startswith('/'):
-                return log_highlighter(obj)
+                return highlight(obj)
             else:
                 return Text(obj)
         else: