+++ pdfalyzer/decorators/pdf_tree_verifier.py
@@ -5,17 +5,14 @@ from dataclasses import dataclass, field
 from types import NoneType
 from typing import Callable, cast
 
-from pypdf import PdfReader
-from pypdf.errors import PdfReadError
 from pypdf.generic import (ArrayObject, BooleanObject, DictionaryObject, IndirectObject, NameObject, NullObject,
      NumberObject, PdfObject, StreamObject)
 from rich.markup import escape
 from yaralyzer.util.logging import log
 
 from pdfalyzer.decorators.document_model_printer import highlighted_raw_pdf_obj_str
-from pdfalyzer.decorators.pdf_tree_node import PdfTreeNode
-from pdfalyzer.helpers.pdf_object_helper import describe_obj
 from pdfalyzer.util.adobe_strings import *
+from pdfalyzer.util.helpers.pdf_object_helper import describe_obj
 
 NUM_PREVIEW_BYTES = 1_024
 OK_UNPLACED_TYPES = (BooleanObject, NameObject, NoneType, NullObject, NumberObject)