+++ pdfalyzer/output/layout.py
@@ -8,7 +8,7 @@ from rich.table import Table
 from yaralyzer.output.console import console, console_width
 from yaralyzer.util.helpers.rich_helper import DEFAULT_TABLE_OPTIONS
 
-from pdfalyzer.helpers.rich_text_helper import indent_padding
+from pdfalyzer.util.helpers.rich_text_helper import indent_padding
 
 DEFAULT_SUBTABLE_COL_STYLES = ['white', 'bright_white']
 HEADER_PADDING = (1, 1)